toolkit/content/tests/chrome/test_preferences_clearUserPref.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test Preference.value = undefined behavior</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="chrome://global/content/preferencesBindings.js"></script>
  <script>
    /* import-globals-from /toolkit/content/preferencesBindings.js */
    function setupPref(prefDef) {
      const defaultBranch = Services.prefs.getDefaultBranch("");

      switch (prefDef.type) {
        case "bool":
          defaultBranch.setBoolPref(prefDef.id, prefDef.defaultValue);
          break;
        case "int":
          defaultBranch.setIntPref(prefDef.id, prefDef.defaultValue);
          break;
        case "string":
          defaultBranch.setStringPref(prefDef.id, prefDef.defaultValue);
          break;
      }

      return Preferences.add({ id: prefDef.id, type: prefDef.type });
    }

    function withNonInstantApply(testFn) {
      const originalType = document.documentElement.getAttribute("type");

      document.documentElement.setAttribute("type", "child");
      is(Preferences.instantApply, false, "non-instantApply mode enabled");

      try {
        testFn();
      } finally {
        if (originalType) {
          document.documentElement.setAttribute("type", originalType);
        } else {
          document.documentElement.removeAttribute("type");
        }
      }
    }

    function watchChanges(pref) {
      let changeCount = 0;
      const listener = () => changeCount++;
      pref.on("change", listener);

      return {
        get count() { return changeCount; },
        cleanup() { pref.off("change", listener); }
      };
    }

    add_task(async function testClearUserValue() {
      const pref = setupPref({ id: "test.pref.bool.one", type: "bool", defaultValue: true });
      const watcher = watchChanges(pref);

      is(pref.value, true, "starts at default");
      is(pref.hasUserValue, false, "no user value");

      // Clearing when already at default is a no-op
      pref.value = undefined;
      is(pref.value, true, "still at default");
      is(pref.hasUserValue, false, "still no user value");
      is(watcher.count, 0, "no change event when clearing default");

      // Set user value and clear it
      pref.value = false;
      is(pref.hasUserValue, true, "user value set");
      is(watcher.count, 1, "change event on value change");

      pref.value = undefined;
      is(pref.value, true, "back to default");
      is(pref.hasUserValue, false, "user value cleared");
      is(watcher.count, 2, "change event on clear to default");

      // Multiple clears work correctly
      pref.value = undefined;
      is(pref.value, true, "remains at default");
      is(watcher.count, 2, "no change event on repeated clear");

      // Verify internal consistency
      is(pref.value, pref.valueFromPreferences, "value equals valueFromPreferences");
      is(pref._value, true, "internal _value correct");

      watcher.cleanup();
    });

    add_task(async function testClearUserValueInt() {
      const pref = setupPref({ id: "test.pref.int.one", type: "int", defaultValue: -1 });
      const watcher = watchChanges(pref);

      is(pref.value, -1, "starts at default");
      is(pref.hasUserValue, false, "no user value");

      // Clearing when already at default is a no-op
      pref.value = undefined;
      is(pref.value, -1, "still at default");
      is(pref.hasUserValue, false, "still no user value");
      is(watcher.count, 0, "no change event when clearing default");

      // Set user value and clear it
      pref.value = 5;
      is(pref.hasUserValue, true, "user value set");
      is(watcher.count, 1, "change event on value change");

      pref.value = undefined;
      is(pref.value, -1, "back to default");
      is(pref.hasUserValue, false, "user value cleared");
      is(watcher.count, 2, "change event on clear to default");

      // Multiple clears work correctly
      pref.value = undefined;
      is(pref.value, -1, "remains at default");
      is(watcher.count, 2, "no change event on repeated clear");

      // Verify internal consistency
      is(pref.value, pref.valueFromPreferences, "value equals valueFromPreferences");
      is(pref._value, -1, "internal _value correct");

      watcher.cleanup();
    });

    add_task(async function testNonInstantApplyMode() {
      withNonInstantApply(() => {
        const pref = setupPref({ id: "test.pref.bool.two", type: "bool", defaultValue: true });
        const watcher = watchChanges(pref);

        pref.value = false;
        is(pref._value, false, "value stored locally");

        pref.value = undefined;
        is(watcher.count, 2, "change events fired");
        is(pref._value, undefined, "value is undefined");

        pref.value = undefined;
        is(watcher.count, 2, "no additional change event");

        watcher.cleanup();
      });
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
